dev: Move cell loaders into `dev::util`
